move platformtype func to manager utils to reuse it.
