+++ pkg/controller/utils/platform.go
@@ -1,0 +1,43 @@
+package utils
+
+import (
+	"context"
+
+	configv1 "github.com/openshift/api/config/v1"
+	"k8s.io/apimachinery/pkg/types"
+	"k8s.io/client-go/tools/clientcmd"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+	"sigs.k8s.io/controller-runtime/pkg/manager"
+)
+
+// PlatformType queries the ku8s api for the infrastructure config map and retrieves the current platform.
+func PlatformType(m manager.Manager) (configv1.PlatformType, error) {
+	client, err := getClient()
+	if err != nil {
+		return configv1.NonePlatformType, err
+	}
+	infraName := types.NamespacedName{Name: "cluster"}
+	infra := &configv1.Infrastructure{}
+	err = client.Get(context.Background(), infraName, infra)
+	if err != nil {
+		return configv1.NonePlatformType, err
+	}
+	return infra.Status.Platform, nil
+}
+
+func getClient() (client.Client, error) {
+	rules := clientcmd.NewDefaultClientConfigLoadingRules()
+	kubeconfig := clientcmd.NewNonInteractiveDeferredLoadingClientConfig(rules, &clientcmd.ConfigOverrides{})
+	cfg, err := kubeconfig.ClientConfig()
+	if err != nil {
+		return nil, err
+	}
+
+	//apis.AddToScheme(scheme.Scheme)
+	dynamicClient, err := client.New(cfg, client.Options{})
+	if err != nil {
+		return nil, err
+	}
+
+	return dynamicClient, nil
+}
